Implemented logic for the apis
- Updated exception handler
- Removed id field from SalaryInfo
- Implemented Repository function 
- Implement util class to read csv file
- Implemented Upload and get api logic
- Updated Data seeding script
- Updated test csv file data

+++ src/main/java/com/dave/salinfo/bean/SalaryInfoBean.java
@@ -10,6 +10,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class SalaryInfoBean {
-    private String name;
-    private double salary;
+  private String name;
+  private float salary;
 }
